Rename enum files
