+++ test-graphics/TestGraphics.cpp
@@ -1,6 +1,6 @@
 #include "TestGraphics.h"
 
-#include <NAS2D/EventHandlerKeyCode.h>
+#include <NAS2D/EnumKeyCode.h>
 #include <NAS2D/Utility.h>
 #include <NAS2D/Filesystem.h>
 #include <NAS2D/Renderer/Renderer.h>
